requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
Django==1.10.3
django-autofixture==0.12.1
django-celery==3.1.17
django-chartjs==1.1
django-compat==1.0.13
django-debug-toolbar==1.6
django-extensions==1.7.3
django-wkhtmltopdf==3.1.0
kombu==4.0.2
psycopg2==2.6.2
python-dateutil==2.6.0
pytz==2016.7
redis==2.10.5
six==1.10.0
sqlparse==0.2.1
uWSGI==2.0.13.1
vine==1.1.3
Werkzeug==0.11.11
XlsxWriter==0.9.3
